#418 🧊 #416 🧊 moved target and targetSymbol to separate file
