+++ packages/core/src/utils/helpers/isTarget.ts
@@ -1,6 +1,6 @@
 import type { HookTarget } from './getElement';
 
-import { targetSymbol } from './getElement';
+import { targetSymbol } from './targetSymbol';
 
 export const isTarget = (target: HookTarget) =>
   typeof target === 'object' && ('current' in target || target.type === targetSymbol);
