+++ packages/core/src/utils/helpers/targetSymbol.ts
@@ -1,0 +1,1 @@
+export const targetSymbol = Symbol('target');
